LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-description-attributes-timing.https-expected.txt


PASS pendingLocalDescription is surfaced at the right time
PASS pendingRemoteDescription is surfaced at the right time
PASS currentLocalDescription is surfaced at the right time
PASS currentRemoteDescription is surfaced at the right time
